style(tools): 工具代码类型转换规范化
- 140+工具文件批量替换C风格cast为static_cast

- 代码风格一致性调整提升类型安全性

+++ extensions/src/runtime/bridge.hpp
@@ -37,7 +37,7 @@ private:
     godot::Ref<godot::StreamPeerTCP> tcp_;
     int port_ = 9601;
     int64_t next_id_ = 1;
-    unsigned long long connecting_since_ = 0;
+    uint64_t connecting_since_ = 0;
 };
 
 } // namespace godot_mcp
